pyTest: split reader tests to testVideo/AudioReader.py
